+++ sync-for-payables/src/sdk/models/_speakeasy_hoisted.ts
@@ -1,0 +1,9 @@
+/* tslint:disable */
+/* eslint-disable */
+
+import type * as shared from './index';
+
+
+export interface ConnectionInfo {}
+
+export interface Tags {}